ringbuf_consume: prevent from the partial "message" reads.
It was possible that ringbuf_consume() would return fewer bytes than it
was produced by ringbuf_produce().  Technically, this can be considered
as not a bug, but there are merits in ensuring that ringbuf_consume()
would consume ranges at the produced "message" boundaries, e.g. ringbuf
could be used as a message queue (think of datagrams).  The fix is to
revent the ABA problem on observed 'ready' (i.e. the 'seen' offset).