Rework the registration mechanism and API.
Store the worker structure in the ringbuf_t as an array and accommodate
the use of this ringbuf mechanism with the shared memory objects.  Now,
ringbuf_setup() takes the number of workers and ringbuf_register() takes
a worker index.  Also, add ringbuf_unregister(), so that the workers can
be gracefully removed.