Now workers are tracked in used/free stacks, and ringbuf_consume() only iterates through in-use workers.
Previously, in-use workers were tracked with a "registered" flag in
ringbuf_worker_t, and ringbuf_consume() iterated through all possible
workers. This could be inefficient if there were far more potential
workers than actual workers.

Presently, ringbuf_worker_t.next is a pointer. If the ABA Problem is
considered to be important, it can be changed to an offset into
ringbuf_t.workers[] and a counter.