Now it passes the tests and the stress-test.
The downside is that failed calls to ringbuf_acquire() consume a
worker-record, and ringbuf_consume() must be called to free up
worker-records on the used-workers stack that are no longer in use.
I'm pretty sure it's not safe to try to remove used-workers from the
stack except in ringbuf_consume(). This probably isn't a problem in
practice, but I'm still not happy about it.