Implemented a queue of worker-records. This performs much better than the stack of worker-records, while accomplishing the same goal. ringbuf_register() can still be called in this variant, i.e. it allows for both permanent registration of worker-records as well as temporary registration.
I think the version-number in the first-used/free-worker indices are
unnecessary if both indices are tracked; I also realize that the use
of a first-used-worker index interferes with mixed permanent and
temporary registration, but then the version-number in the
worker-indices must be kept.
Both variants will be tried in further commits.